fix dtype

+++ src/aspire/covariance/covar2d.py
@@ -596,7 +596,7 @@ class BatchedRotCov2D(RotCov2D):
         M_covar = BlkDiagMatrix.zeros_like(A_mean)
 
         for k in np.unique(ctf_idx):
-            weight = np.count_nonzero(ctf_idx == k) / src.n
+            weight = (np.count_nonzero(ctf_idx == k) / src.n).astype(self.dtype)
 
             ctf_basis_k = ctf_basis[k]
             ctf_basis_k_t = ctf_basis_k.T
